fix: move Card component to Card folder
